add LivePlotter class
